\name{GetExpectation}
\alias{GetExpectation}
\title{Returns expected utility}
\description{
The function returns float value of expected utility.
}
\usage{
GetExpectation(Limid)
}
\arguments{
\item{Limid}{LIMID, which has been already created in the workspace. The action is carried out on this network. }
}
\examples{
exp <- GetExpectation(id)
}
\keyword{GetExpectation}