\name{GetLag}
\alias{GetLag}
\title{Gets the number of lag, which was defined for FixLagSmoothing prosedure.}
\description{
The function returns the number of lag for FixLagSmoothing prosedure.
 }
\usage{
GetLag(DBN)
}
\arguments{
\item{DBN}{DBN, which has been already created in the workspace. The action is carried out on this network. }
}
\keyword{GetLag}