\name{GetNumberOfCliques}
\alias{GetNumberOfCliques}
\title{Returns number of cliques}
\description{
The function returns number of cliques in network
}
\usage{
GetNumberOfCliques(MRF) 
}
\arguments{
\item{MRF}{MRF, which has been already created in the workspace. The action is carried out on this network. }
}
\examples{
NC <- GetNumberOfCliques(mrf) 
}
\keyword{GetNumberOfCliques}