\name{GetNumSlices}
\alias{GetNumSlices}
\title{Gets number of slices in the network}
\description{
The function gets number of network slices.
 }
\usage{
GetNumSlices(DBN)
}
\arguments{
\item{DBN}{DBN, which has been already created in the workspace. The action is carried out on this network. }
}
\examples{
##Getting the number of slices : 
NumSlices <- GetNumSlices(dbn) 
}
\keyword{SetNumSlices}