\name{GetPolitics}
\alias{GetPolitics}
\title{Returns pure politics}
\description{
The function returns pure politics as vector of strings.
}
\usage{
GetPolitics(Limid)
}
\arguments{
\item{Limid}{LIMID, which has been already created in the workspace. The action is carried out on this network. }
}
\examples{
politics <- GetPolitics(id)
}
\keyword{GetPolitics}