\name{IsFullDBN}
\alias{IsFullDBN}
\title{Checks the DBN completeness}
\description{
Indicates, wheather it is possibló to construct DBN on the base of specified nodes and edges
 }
\usage{
IsFullDBN(DBN)
}
\arguments{
\item{DBN}{DBN, which has been already created in the workspace. The action is carried out on this network. }
}
\keyword{IsFullDBN}